Murmurhash3: fix big-endian
This modifies the getblock() helpers to produce the same result regardless
of endianess.

Tested for s390x Linux, making sure that the result is now the same
as on other architectures, and by looking at the object code to ensure
that it uses the most efficient instructions (load reversed LRV/LRVG)
for accessing the input.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>