Murmurhash2: make it endian neutral
Use the same logic as Murmurhash3 to load little-endian data from
the input key in all variants of Murmurhash2. The resulting object
code is unchanged on x86, and it does the correct thing on s390,
producing the same result while using the load-reverse instructions.

This is still broken on architectures that cannot handle unaligned
input, but those are becoming very rare.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>